refactor(error-boundary): extract fallback into sibling component file
`app-error-boundary.tsx` was 193 lines, nearly 2× the documented
100-line ceiling, and the comment block describing `role="alert"` /
`aria-live` / focus management sat directly above an unrelated
`SHORT_ID_LEN` constant, 40 lines from the JSX it documented.

Extract `Fallback` and its helpers (`formatTraceId`, `toErrorShape`,
`SHORT_ID_LEN`) into `fallback.component.tsx` as a named export. The
accessibility comment now sits immediately above the
`<div role="alert">` it describes. `app-error-boundary.tsx` is now
50 lines and contains only the wrapper. The `<AppErrorBoundary>` public
API is unchanged; `app-error-boundary.spec.tsx` passes as-is.
